Fix an error where extension broke https://e.mail.ru/inbox
The error was happenning because extension normalized arguments of define, to be more precise, it always made sure that there were 3 arguments, and the second one (dependencies) is array. But requirejs allows modules to skip dependencies if they're only using helpers from require. It does that based on the second argument to define and arity of the callback but because the second argument was always an array, it never did that. So, the error was fixed by passing correct argument to define to allow it to inject helpers.